fix auth bug where jwt was malforme
