+++ src/Pages/Authorize/Authorize.scss
@@ -1,0 +1,3 @@
+.authorize-wrapper {
+  margin-top:5em;
+}
