+++ src/layouts/Navbar/MobileNav.js
@@ -38,7 +38,7 @@ const MobileNav = inject("UserStore")(observer(({redditProfile, UserStore, exten
             {!UserStore.getUser() && 
               <React.Fragment>
                 <li className="d-f ai-c nav-link ">
-                  <Link onClick={setExtended} to="/signup" >Sign Up</Link>
+                  <Link onClick={setExtended} to="/authorize" >Sign Up</Link>
                 </li>
                 <li className="d-f ai-c nav-link">
                   <Link onClick={setExtended} to="/login" >Login</Link>
